backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.3.0
PyYAML==6.0
django-environ==0.10.0
gunicorn==20.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
django-filter==21.1
fpdf
